Adds CPR validation to normal donation widget
